[#38] Move common function to helpers for preset
