+++ scenarios/preset/check_objects_in_preset.py
@@ -2,13 +2,12 @@
 
 import argparse
 import json
-import shlex
-from subprocess import check_output, CalledProcessError, STDOUT
+
+from helpers.neofs_cli import get_object
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--endpoint', help='Node address')
 parser.add_argument('--preset_json', help='JSON file path with preset')
-parser.add_argument('--print_success', help='Print objects that was successfully read', default=False)
 
 args = parser.parse_args()
 
@@ -26,35 +25,15 @@ def main():
         oid = obj.get('object')
         cid = obj.get('container')
 
-        cmd_line = f"neofs-cli object get -r {args.endpoint} -g" \
-                   f" --cid {cid} --oid {oid} --file /dev/null"
+        rst = get_object(cid, oid, args.endpoint, "/dev/null")
 
-        output, success = execute_cmd(cmd_line)
-
-        if success:
+        if rst:
             success_objs += 1
-            if args.print_success:
-                print(f'Object: {oid} from {cid}: {"Ok" if success else "False"}')
         else:
             failed_objs += 1
-            print(f'Object: {oid} from {cid}: {"Ok" if success else "False"}')
 
     print(f'Success objects: {success_objs}')
     print(f'Failed objects: {failed_objs}')
-
-
-def execute_cmd(cmd_line):
-    cmd_args = shlex.split(cmd_line)
-
-    try:
-        output = check_output(cmd_args, stderr=STDOUT).decode()
-        success = True
-
-    except CalledProcessError as e:
-        output = e.output.decode()
-        success = False
-
-    return output, success
 
 
 if __name__ == "__main__":
